requirements.txt
streamlit
tensorflow==2.13
tf-explain
opencv-python
matplotlib
pillow